+++ app/v1/modules/kubernetes_services/__init__.py
@@ -1,19 +1,21 @@
 # -*- coding: utf-8 -*-
 
 
-from .script_read_annotations import ScriptAnnotationsReaderservice
+from .script_read_annotations import ScriptAnnotationsReaderService
 from .script_to_dict import ScriptToDictParserService
 from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
 from .script_read_env_from import ScriptEnvironReaderFromService
+from .script_read_ingress_rules import ScriptIngressRulesReaderService
 from .script_read_volumes_mounts import ScriptVolumeMountsReaderService
 
 
 __all__ = [
-    "ScriptAnnotationsReaderservice",
+    "ScriptAnnotationsReaderService",
     "ScriptToDictParserService",
     "ScriptVolumesReaderService",
     "ScriptEnvironReaderFromService",
     "ScriptEnvironReaderService",
+    "ScriptIngressRulesReaderService",
     "ScriptVolumeMountsReaderService",
 ]
