+++ app/v1/modules/kubernetes_services/script_read_annotations.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 
-class ScriptAnnotationsReaderservice:
+class ScriptAnnotationsReaderService:
     """
-    ScriptAnnotationsReaderservice
+    ScriptAnnotationsReaderService
     """
 
     @staticmethod
